+++ src/components/types/mutextedOrderMap/mutextedOrderMap.go
@@ -1,0 +1,11 @@
+package mutextedOrderMap
+
+import (
+	"kitchen/src/components/types/order"
+	"sync"
+)
+
+type MutextedOrderMap struct {
+	OrderMap 	map[string]order.Order
+	Mutex 	  *sync.Mutex
+}
