+++ src/components/types/mutextedValue/mutextedValue.go
@@ -4,5 +4,5 @@ import "sync"
 
 type MutextedValue struct {
 	Value 	int
-	Mutex 	sync.Mutex
+	Mutex 	*sync.Mutex
 }
